# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TaskTrollolo API",
    "description": "\n\n API documentation for TaskTrollolo - Our final project \n\n💾 [Backend Repository](https://github.com/exigibil/Tasktrollolo-backend) - \n\n🖥 [Frontend Repository](https://github.com/exigibil/Tasktrollolo) \n\n📂 [Figma Design](https://www.figma.com/design/fJF13s2UlxPIwTMcPVrSiz/TaskPro?t=BRwoZsq1XJGXwShy-0) \n\n This project was developed by: \n\n1.[Adrian Popa - Team Leader](https://github.com/exigibil), \n\n2.[Florentina Vachente - Scrum Master](https://github.com/FloriGOIT), \n\n3.[Gavrila Daniela](https://github.com/gavrila0909), \n\n4.[Ovidiu Bodean](https://github.com/ovidutzul), \n\n5.[Andrei-Gabriel Paiu-Rotundu](https://github.com/djxsrouk)"
  },
  "host": "localhost:2000",
  "basePath": "/auth",
  "servers": ["http"],
...
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Registers a new user with name, email, and password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["name", "email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Registration successful"
          },
          "409": {
            "description": "Email already in use"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login a user",
        "description": "Logs in a user with email and password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Incorrect email or password"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout a user",
        "description": "Logs out the authenticated user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}